Move compression.zstd.zstdfile to compression.zstd._zstdfile
This makes it private.
